Add filter and search tiles.
