+++ src/collective/collectionfilter/tiles/search.pt
@@ -1,4 +1,4 @@
-<aside class="portlet ${view/filterClassName} portletCollectionSearch pat-collectionfilter"
+<aside class="collectionSearch ${view/filterClassName} pat-collectionfilter"
        data-pat-collectionfilter='{
            "collectionUUID": "${view/data/target_collection}",
            "reloadURL": "${view/reload_url}"
@@ -6,11 +6,8 @@
        tal:condition="view/available"
        i18n:domain="collective.collectionfilter">
 
-  <header class="portletHeader"
-          tal:content="view/title">
-          Title
-  </header>
-  <div class="portletContent">
+  <header tal:content="view/title">Title</header>
+  <div class="searchContent">
     <form method="get" name="searchForm" role="form"
         tal:attributes="action view/action_url">
         <tal:urlquery repeat="item python:view.urlquery.items()">
